Started to implement the Plug-in gait
